Refactor peak observation functions into separate file
